Build fixes.
-Werror was added to the default config, so turn that off
for most of mesa and fix the few warnings that aren't
covered by that.

+++ src/egl/main/eglglobals.c
@@ -82,7 +82,7 @@ _eglAddAtExitCall(void (*func)(void))
          registered = EGL_TRUE;
       }
 
-      assert(_eglGlobal.NumAtExitCalls < ARRAY_SIZE(_eglGlobal.AtExitCalls));
+      assert((unsigned int)_eglGlobal.NumAtExitCalls < ARRAY_SIZE(_eglGlobal.AtExitCalls));
       _eglGlobal.AtExitCalls[_eglGlobal.NumAtExitCalls++] = func;
 
       mtx_unlock(_eglGlobal.Mutex);
